better organization of files (moved inferences to their own file)
