+++ inference_test.go
@@ -2,8 +2,11 @@ package groq_test
 
 import (
 	"bytes"
+	"context"
+	"encoding/json"
 	"fmt"
 	"io"
+	"net/http"
 	"os"
 	"path/filepath"
 	"testing"
@@ -11,8 +14,60 @@ import (
 	groq "github.com/conneroisu/groq-go"
 	"github.com/conneroisu/groq-go/pkg/builders"
 	"github.com/conneroisu/groq-go/pkg/test"
+	"github.com/conneroisu/groq-go/pkg/tools"
 	"github.com/stretchr/testify/assert"
 )
+
+func TestChat(t *testing.T) {
+	ctx := context.Background()
+	a := assert.New(t)
+	ts := test.NewTestServer()
+	returnObj := groq.ChatCompletionResponse{
+		ID:      "chatcmpl-123",
+		Object:  "chat.completion.chunk",
+		Created: 1693721698,
+		Model:   "llama3-groq-70b-8192-tool-use-preview",
+		Choices: []groq.ChatCompletionChoice{
+			{
+				Index: 0,
+				Message: groq.ChatCompletionMessage{
+					Role:    groq.RoleAssistant,
+					Content: "Hello!",
+				},
+			},
+		},
+	}
+	ts.RegisterHandler("/v1/chat/completions", func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusOK)
+		jsval, err := json.Marshal(returnObj)
+		a.NoError(err)
+		_, err = w.Write(jsval)
+		if err != nil {
+			t.Fatal(err)
+		}
+	})
+	testS := ts.GroqTestServer()
+	testS.Start()
+	client, err := groq.NewClient(
+		test.GetTestToken(),
+		groq.WithBaseURL(testS.URL+"/v1"),
+	)
+	a.NoError(err)
+	resp, err := client.ChatCompletion(ctx, groq.ChatCompletionRequest{
+		Model: groq.ModelLlama3Groq70B8192ToolUsePreview,
+		Messages: []groq.ChatCompletionMessage{
+			{
+				Role:    groq.RoleUser,
+				Content: "Hello!",
+			},
+		},
+		MaxTokens: 2000,
+		Tools:     []tools.Tool{},
+	})
+	a.NoError(err)
+	a.NotEmpty(resp.Choices[0].Message.Content)
+}
 
 func TestAudioWithFailingFormBuilder(t *testing.T) {
 	a := assert.New(t)
@@ -77,6 +132,24 @@ func TestAudioWithFailingFormBuilder(t *testing.T) {
 	}
 }
 
+func TestModeration(t *testing.T) {
+	a := assert.New(t)
+	ctx := context.Background()
+	client, server, teardown := setupGroqTestServer()
+	defer teardown()
+	server.RegisterHandler("/v1/chat/completions", handleModerationEndpoint)
+	mod, err := client.Moderate(ctx,
+		[]groq.ChatCompletionMessage{
+			{
+				Role:    groq.RoleUser,
+				Content: "I want to kill them.",
+			},
+		},
+		groq.ModelLlamaGuard38B,
+	)
+	a.NoError(err)
+	a.NotEmpty(mod)
+}
 func TestCreateFileField(t *testing.T) {
 	a := assert.New(t)
 	t.Run("createFileField failing file", func(t *testing.T) {
